Move script types to primitives
Move the `Script` and `ScriptBuf` types to `primitives`, nothing else.
