+++ primitives/Cargo.toml
@@ -30,7 +30,7 @@ units = { package = "bitcoin-units", version = "0.1.0", default-features = false
 
 arbitrary = { version = "1", optional = true }
 ordered = { version = "0.2.0", optional = true }
-serde = { version = "1.0.103", default-features = false, features = ["derive"], optional = true }
+serde = { version = "1.0.103", default-features = false, features = ["derive", "alloc"], optional = true }
 
 [dev-dependencies]
